# RealworldWeb/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "RealWorld Conduit API",
    "description" : "Conduit API documentation",
    "contact" : {
      "name" : "RealWorld",
      "url" : "https://realworld.how"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://api.realworld.io/api"
  } ],
...
  "paths" : {
...
    "/users" : {
      "post" : {
        "tags" : [ "User and Authentication" ],
        "description" : "Register a new user",
        "operationId" : "CreateUser",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/NewUserRequest"
        },
        "responses" : {
          "201" : {
            "description" : "User",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name" : "body"
      }
    },
    "/user" : {
      "get" : {
        "tags" : [ "User and Authentication" ],
        "summary" : "Get current user",
        "description" : "Gets the currently logged-in user",
        "operationId" : "GetCurrentUser",
        "responses" : {
          "200" : {
            "description" : "User",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "User and Authentication" ],
        "summary" : "Update current user",
        "description" : "Updated user information for current user",
        "operationId" : "UpdateCurrentUser",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/UpdateUserRequest"
        },
        "responses" : {
          "200" : {
            "description" : "User",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "content" : { }
          },
          "422" : {
            "description" : "Unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GenericErrorModel"
                }
              }
            }
          }
        },
        "security" : [ {
          "Token" : [ ]
        } ],
        "x-codegen-request-body-name" : "body"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "NewUser" : {
        "required" : [ "email", "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string",
            "format" : "password"
          }
        }
      },
      "User" : {
        "required" : [ "bio", "email", "image", "token", "username" ],
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          }
        },
        "example" : {
          "image" : "image",
          "bio" : "bio",
          "email" : "email",
          "token" : "token",
          "username" : "username"
        }
      },
      "UpdateUser" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          },
          "username" : {
            "type" : "string"
          },
          "bio" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          }
        }
      },
...
      "GenericErrorModel" : {
        "required" : [ "errors" ],
        "type" : "object",
        "properties" : {
          "errors" : {
            "$ref" : "#/components/schemas/GenericErrorModel_errors"
          }
        }
      },
      "inline_response_200" : {
        "required" : [ "user" ],
        "type" : "object",
        "properties" : {
          "user" : {
            "$ref" : "#/components/schemas/User"
          }
        },
        "example" : {
          "user" : {
            "image" : "image",
            "bio" : "bio",
            "email" : "email",
            "token" : "token",
            "username" : "username"
          }
        }
      },
...
      "GenericErrorModel_errors" : {
        "required" : [ "body" ],
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "requestBodies" : {
...
      "NewUserRequest" : {
        "description" : "Details of the new user to register",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "user" ],
              "type" : "object",
              "properties" : {
                "user" : {
                  "$ref" : "#/components/schemas/NewUser"
                }
              }
            }
          }
        },
        "required" : true
      },
      "UpdateUserRequest" : {
        "description" : "User details to update. At least **one** field is required.",
        "content" : {
          "application/json" : {
            "schema" : {
              "required" : [ "user" ],
              "type" : "object",
              "properties" : {
                "user" : {
                  "$ref" : "#/components/schemas/UpdateUser"
                }
              }
            }
          }
        },
        "required" : true
      },
...
    },
    "securitySchemes" : {
      "Token" : {
        "type" : "apiKey",
        "description" : "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "name" : "Authorization",
        "in" : "header"
      }
    }
  }
}